Move tests for `remove` into separate file
