+++ test/operations-unit.js
@@ -22,46 +22,6 @@ describe('operations', () => {
   })
 
 
-  describe('remove items from selection', () => {
-    let newState = null
-
-    beforeEach(() => {
-      const items = iterable(['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H'])
-      const state = init()
-
-      newState = flow(
-        setItems(items),
-        setSelection(['A', 'B', 'C', 'D', 'F']),
-        remove(['A', 'D', 'B'])
-      )(state)
-    })
-
-    it('should remove items from selection', () => {
-      expectExactlySameMembers(getSelection(newState), ['C', 'F'])
-    })
-
-    it('should return items as newly deselected', () => {
-      expectExactlySameMembers(getChangedDeselection(newState), ['A', 'D', 'B'])
-    })
-
-    it('should return empty list as newly selected', () => {
-      expectExactlySameMembers(getChangedSelection(newState), [])
-    })
-
-    it('should return only previously selected items as newly deselected', () => {
-      const items = iterable(['A', 'B', 'C', 'D'])
-      const state = init()
-
-      newState = flow(
-        setItems(items),
-        setSelection(['A', 'B']),
-        remove(['A', 'B', 'C'])
-      )(state)
-
-      expectExactlySameMembers(getChangedDeselection(newState), ['A', 'B'])
-    })
-  })
-
   describe('remove all items from selection', () => {
     let newState = null
